ProjectZen initial commit, exploring a new 'proper' Java string class, and more
Duplicated with simple modifications the contributors agreement from
the StringTemplate and ANTLR projects, in contributors.txt.

The GNU LGPL advises electronic and paper mail contact information so
these have been provided in contact.txt.

Source files are in src/ and I intend this project to collect my bash/
shell library in time, so the Java stuff is in src/java/* and in time
the bash shell stuff will be in src/sh/ for example.

The very first draft of zen.lang.string has been added to
src/java/zen/lang/string.java - although a good week or more has been
spent here and there reading, researching and collating the findings
from this activity to string.java, and I checked that it compiles.