Javadoc +Swift, unpack surrogates, grapheme clusters
Swift is the first mainstream language I am aware of that handles Unicode
characters as they should be - as grapheme clusters; added a few links and
references to this.

Unpacked surrogates a little for more clarity plus references to the Unicode
FAQ, once again it's clear to me.

Added a little internal cross referencing in regard to above.

Added urls.txt in anticipation of pointing to other git repos/ clones, and to
the canonical Javadoc.

Added bin/javadoc.sh with the command I use to generate the Javadocs.

Enhanced contributors.txt